+++ beacon-chain/node/config.go
@@ -79,30 +79,6 @@ func configureBuilderCircuitBreaker(cliCtx *cli.Context) error {
 	return nil
 }
 
-// configureBlobRetentionEpoch sets the for blob retention based on command-line context. It overrides the existing network configuration
-// for blob retention if the input epoch is greater than or equal to the spec default value.
-// An error if the input epoch is smaller than the spec default value.
-func configureBlobRetentionEpoch(cliCtx *cli.Context) error {
-	// Check if the blob retention epoch flag is set.
-	if cliCtx.IsSet(flags.BlobRetentionEpoch.Name) {
-		// Retrieve and cast the epoch value.
-		epochValue := cliCtx.Uint64(flags.BlobRetentionEpoch.Name)
-		e := primitives.Epoch(epochValue)
-
-		// Validate the epoch value against the spec default.
-		if e < blobRetentionEpochSpecValue {
-			return fmt.Errorf("input blob retention epoch smaller than spec default, %d < %d", e, blobRetentionEpochSpecValue)
-		}
-
-		// Copy the existing beacon network configuration.
-		c := params.BeaconNetworkConfig().Copy()
-		c.MinEpochsForBlobsSidecarsRequest = e
-		params.OverrideBeaconNetworkConfig(c)
-	}
-
-	return nil
-}
-
 func configureSlotsPerArchivedPoint(cliCtx *cli.Context) error {
 	if cliCtx.IsSet(flags.SlotsPerArchivedPoint.Name) {
 		c := params.BeaconConfig().Copy()
